split method - more accurate description
At first I did not understand the meaning of this method from the documentation, and at first I thought that the address was simply broken into pieces of exponent bits. But looking at the code, I realized that it is generally intended for another. Apparently this method was used to divide the networks into subnets a exponent bits additionally? Then my comment more accurately reflects the meaning IMHO.